Fix resume functionality.

+++ library/src/main/java/com/novoda/downloadmanager/download/task/State.java
@@ -2,10 +2,10 @@ package com.novoda.downloadmanager.download.task;
 
 class State {
     public boolean gotData;
-    public int currentBytes;
+    public long currentBytes;
     public long speedSampleStart;
-    public int speedSampleBytes;
+    public long speedSampleBytes;
     public long speed;
-    public int bytesNotified;
+    public long bytesNotified;
     public long timeLastNotification;
 }
